Small update

+++ src/AcesDurables.tsx
@@ -3,14 +3,12 @@ import { D1Database } from "./d1_beta";
 import { Env } from "./env";
 import { getSessionUser, TenantSessionUser } from "./session";
 import {
-  Admin,
   Backupable,
   Client,
   COLUMNS,
   DurableType,
   getDurableType,
   PREFIX,
-  Prefixes,
   Project,
   TableName,
   TableNames,
@@ -26,7 +24,6 @@ import {
   remove_duplicates_es6,
 } from "./utils";
 import { aces_sample_data } from "../sample-data/do-ready";
-import { rebuildSeedKV } from "./dev-routes";
 import { dumpItem, prepItemForDb } from "./store";
 
 function singular(tableName: TableName) {
@@ -85,56 +82,16 @@ export class AcesDurables {
     // PUT ALL INITIATION HERE
     this.state.blockConcurrencyWhile(async () => {
       console.log("blockConcurrencyWhile( ... )");
-      // if (!(await this.storage.get(KEYS))) {
-      //   console.log("Preparing...");
-      //   const step1 = await this.storage.put(KEYS, [KEYS]);
-      //   console.log(step1);
-      //   const step2 = await this.storage.put(aces_sample_data);
-      //   console.log(step2);
-      //   const keys = Object.keys(aces_sample_data);
-      //   console.log("Saving keys...");
-      //   const step3 = await this.storage.put(KEYS, keys);
-      //   console.log("Finished preparation");
-      // }
-      this.initStoreWithSampleData();
-      console.log("Finished initiation");
-    });
-
-    this.app.get("/do-weep", async (c) => {
-      // delete(keys, options)
-      // =====================
-      // Deletes the provided keys and their associated values.
-      // Supports up to 128 keys at a time. Returns a count of
-      // the number of key-value pairs deleted.
-      // const list = await this.storage.list({ prefix: "account" });
-      // const list = await this.storage.list(); // <- ERROR: Trace: DurableObjectError [ERR_DESERIALIZATION]: Unable to deserialize stored Durable Object data due to an invalid or unsupported version.
-      // const array: any[] = Array.from(list).map(([key, value]) => key);
-      // const keys = (await this.storage.get(KEYS)) as string[];
-      // const rs = await this.storage.delete(keys);
-      const rs = await this.storage.deleteAll();
-      // const rs = await this.storage.deleteAll()
-      // return c.text("do-weep");
-      return c.json(rs);
-    });
-
-    this.app.get("/do-seed", async (c) => {
-      // const list = await this.storage.list();
-      // if (list) {
-      //   const array: any[] = Array.from(list).map(([key, value]) => key);
-      //   return c.json(array);
-      // }
-      let RS;
-      try {
-        const rs = await this.storage.put(aces_sample_data);
-        RS = rs;
-      } catch (error) {
-        console.log("DO-SEED ERROR");
-        console.log(error);
-        console.log("=====================");
-        return c.text("ERROR");
+      console.log("Checking keys...");
+      const keys = await this.storage.get(KEYS); // as string[];
+      if (keys == undefined || (keys as string[]).length == 0) {
+        console.log("Storage empty");
+        this.initStoreWithSampleData(true);
       }
-      // rs.
-      return c.text(RS ? RS : "do-seed");
+      if (keys && (keys as string[]).length > 1) {
+        console.log(`Found ${(keys as string).length} key(s)`);
+      }
+      console.log("Finished initiation");
     });
 
     /* Return all DurableObject keys */
@@ -156,6 +113,11 @@ export class AcesDurables {
         console.log(error);
         return c.text("DO ERROR 103");
       }
+    });
+
+    this.app.get("/dump", async (c) => {
+      const rs = await this.dumpStorage();
+      return c.text(rs.join("\n"));
     });
 
     this.app.get("/tenant", async (c) => {
@@ -215,15 +177,6 @@ export class AcesDurables {
     this.app.get("/:path", async (c) => {
       const path = c.req.param("path");
       console.log(path);
-      if (path == "delete-db") {
-        // await this.deleteDB(this.env)
-        // await this.saveToDBTable(this.env);
-        const rs = await this.dumpStorage();
-        return c.text(rs.join("\n"));
-        return c.json({
-          info: "After backup to D1",
-        });
-      }
       const url = new URL(c.req.url);
       const group = url.searchParams.get("group");
       const method = url.searchParams.get("method");
@@ -259,15 +212,6 @@ export class AcesDurables {
           500
         );
       }
-
-      // TOBEDELETED [
-      /*
-      const dump: string[] = []
-      results.forEach(item => {
-        dump.push(dumpItem(item, 'accounts'))
-      })
-      return c.text(dump.join("\n")) // */
-      // TOBEDELETED ]
 
       // Case for modules: always return modules with their groups
       if (tableName == "modules") {
@@ -323,11 +267,6 @@ export class AcesDurables {
       const array: any[] = Array.from(list).map(([, value]) => value);
       const result = array[0];
 
-      // TOBEDELETED [
-      // const types = dumpItem(result, 'accounts')
-      // return c.text(types)
-      // TOBEDELETED ]
-
       // Case modules: include group
       if (tableName == "modules") {
         const group = await this.storage.get(
@@ -348,13 +287,6 @@ export class AcesDurables {
       const subpath = c.req.param("subpath");
       return c.text(`query -> ${path} -> ${subpath} -> ${opt}`);
     });
-
-    /* Delete db */
-    // this.app.get("/delete-db", async (c) => {
-    //   return c.json({
-    //     info: "Delete DB",
-    //   });
-    // });
   }
 
   async fetch(request: Request) {
@@ -363,7 +295,7 @@ export class AcesDurables {
     return this.app.fetch(request);
   }
 
-  deleteDB = async (env: Env) => {
+  __deleteDB = async (env: Env) => {
     const rs = await env.DB.batch([
       env.DB.prepare(`DELETE FROM project_modules`),
       env.DB.prepare(`DELETE FROM modules`),
@@ -384,7 +316,7 @@ export class AcesDurables {
    * @param env Env
    * @returns array of results report
    */
-  saveToDBTable = async (env: Env) => {
+  __saveToDBTable = async (env: Env) => {
     const run = async (e, t, c, p) => {
       const list = await this.storage.list({ prefix: p });
       if (list.size == 0) return;
@@ -420,42 +352,44 @@ export class AcesDurables {
    */
   dumpStorage = async () => {
     const run = async (s, p, t: TableName) => {
-      if (ViewNames.includes(t)) return []
+      if (ViewNames.includes(t)) return [];
       const list = await s.list({ prefix: p });
       const array = Array.from(list).map(([, v]) => v);
-      const rs = []
+      const rs = [];
       array.forEach(async (item) => {
         const text = dumpItem(item, t);
         rs.push(text);
       });
-      return rs.length ? rs : ['-- empty']
+      return rs.length ? rs : ["-- empty"];
     };
 
-    const rs = []
+    const rs = [];
     // DO NOT CHANGE THE ORDER
-    rs.push("--\n-- admins\n--")
+    rs.push("--\n-- admins\n--");
     rs.push(...(await run(this.storage, PREFIX.admin, "admins")));
-    rs.push("--\n-- users\n--")
+    rs.push("--\n-- users\n--");
     rs.push(...(await run(this.storage, PREFIX.user, "users")));
-    rs.push("--\n-- tenants\n--")
+    rs.push("--\n-- tenants\n--");
     rs.push(...(await run(this.storage, PREFIX.tenant, "tenants")));
-    rs.push("--\n-- members\n--")
+    rs.push("--\n-- members\n--");
     rs.push(...(await run(this.storage, PREFIX.member, "members")));
-    rs.push("--\n-- clients\n--")
+    rs.push("--\n-- clients\n--");
     rs.push(...(await run(this.storage, PREFIX.client, "clients")));
-    rs.push("--\n-- projects\n--")
+    rs.push("--\n-- projects\n--");
     rs.push(...(await run(this.storage, PREFIX.project, "projects")));
-    rs.push("--\n-- module_groups\n--")
+    rs.push("--\n-- module_groups\n--");
     rs.push(...(await run(this.storage, PREFIX.modulegroup, "module_groups")));
-    rs.push("--\n-- modules\n--")
+    rs.push("--\n-- modules\n--");
     rs.push(...(await run(this.storage, PREFIX.module, "modules")));
-    rs.push("--\n-- project_modules\n--")
-    rs.push(...(await run(this.storage, PREFIX.projectmodule, "project_modules")));
+    rs.push("--\n-- project_modules\n--");
+    rs.push(
+      ...(await run(this.storage, PREFIX.projectmodule, "project_modules"))
+    );
     // rs.push('-- accounts')
     // rs.push(...(await run(this.storage, PREFIX.account, "accounts")));
     // rs.push('-- module_usages')
     // rs.push(...(await run(this.storage, PREFIX.moduleusage, "module_usages")));
-    return rs
+    return rs;
   };
 
   __resetKeys = async () => {
@@ -476,25 +410,27 @@ export class AcesDurables {
     }
   };
 
-  initStoreWithSampleData = async () => {
-    console.log("storage.deleteAll()");
-    await this.storage.deleteAll();
-    // Rebuild sample data
-    console.log("storage.transaction()");
-    await this.storage.transaction(async (txn) => {
-      await txn.put(aces_sample_data); // <- max 128 entries
-      const keys = Object.keys(aces_sample_data);
-      await txn.put(KEYS, keys);
-    });
-    // Reset DB
-    console.log("deleteDB()");
-    // await this.deleteDB();
-    // Rebuild DB
-    console.log("backupToDB()");
-    // await this.backupToDB();
+  initStoreWithSampleData = async (force: boolean) => {
+    if (force) {
+      console.log("Force init...");
+      console.log("Deleting all key-value pairs in storage...");
+      await this.storage.deleteAll();
+      console.log("Loading sample data...");
+      await this.storage.transaction(async (txn) => {
+        await txn.put(aces_sample_data); // <- max 128 entries
+        const keys = Object.keys(aces_sample_data);
+        await txn.put(KEYS, keys);
+      });
+      console.log(
+        `Loaded ${
+          Object.keys(aces_sample_data).length
+        } key-value pairs to storage.`
+      );
+      console.log("Finished");
+    }
   };
 
-  loadFromDB = async (tableName: TableName, force = false) => {
+  __loadFromDB = async (tableName: TableName, force = false) => {
     if (!TableNames.includes(tableName)) return;
 
     const list = await this.storage.list({
@@ -524,6 +460,7 @@ export class AcesDurables {
     }
   };
 
+  // NOT DONE YET
   updateItem = (
     table: string,
     id: string,
@@ -541,6 +478,7 @@ export class AcesDurables {
     return updatedObject;
   };
 
+  // NOT DONE YET
   saveNewItemAndUpdateKeys = async (
     item: Client | Project | Tenant,
     key: string,
